+++ README.md
@@ -5,3 +5,4 @@
 > Take a look at the `content.tsx` file where we have a simple list of countries. Do you think you can set up and connect an external database so that this list is not hard coded? I hear Supabase (https://supabase.com/) is perfect for this type of work.
 
 > When you are done, please submit a PR (pull request) back to this repo and I can review it.
+OQmvwJ9t3Z7D5UIR password for supabase project
